requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
crypto==1.4.1
cycler==0.12.1
exceptiongroup==1.2.1
fonttools==4.51.0
galois==0.3.8
idna==3.7
iniconfig==2.0.0
kiwisolver==1.4.5
llvmlite==0.42.0
matplotlib==3.9.0
mpmath==1.3.0
Naked==0.1.32
nose2==0.14.2
numba==0.59.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
pycryptodome==3.20.0
pyparsing==3.1.2
pytest==8.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
shellescape==3.8.1
six==1.16.0
sympy==1.12
tomli==2.0.1
typing_extensions==4.11.0
urllib3==2.2.1

typer~=0.13.0
pikepdf~=9.4.0
setuptools~=68.2.0
PyMuPDF~=1.24.13
pdfquery~=0.4.3
pdftotext~=2.2.2
pypdf~=5.1.0
xmptools~=0.6.2
pdfplumber~=0.11.4
rlextra~=4.2.5.1
reportlab~=4.2.5
borb~=2.1.25